Add relation between authors and books OK

+++ src/main/java/com/globant/academy/springdata/repositories/AuthorRepository.java
@@ -3,6 +3,6 @@ package com.globant.academy.springdata.repositories;
 import com.globant.academy.springdata.entities.Author;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AuthorRepository extends CrudRepository<Author, Long> {
+public interface AuthorRepository extends CrudRepository<Author, Integer> {
 
 }
