+++ src/main/java/com/globant/academy/springdata/repositories/BookRepository.java
@@ -3,5 +3,5 @@ package com.globant.academy.springdata.repositories;
 import com.globant.academy.springdata.entities.Book;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BookRepository extends CrudRepository<Book,Long> {
+public interface BookRepository extends CrudRepository<Book,Integer> {
 }
